added chat context
The chat program now will be fed the entirety of the chat as a prompt. This gives the bot a "memory" so it can "recall" things said earlier in the conversation. This is limited however by the token limit of the whatever model you are using (prompt will eventually take too many tokens to process)

Added a self chat, not quite working properly. Need to test using two separate "memories" so that one chatbot can "remember" the intro prompt but the other cannot